+++ services/keyvault/mgmt/2021-10-01/keyvault/CHANGELOG.md
@@ -1,2 +1,10 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. MHSMPrivateEndpointConnectionItem.Etag
+1. MHSMPrivateEndpointConnectionItem.ID